server/requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.5
billiard==4.1.0
black==23.3.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
Deprecated==1.2.14
dill==0.3.6
Django==4.2.2
django-celery-beat==2.5.0
django-cors-headers==4.1.0
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flower==1.2.0
humanize==4.6.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
kombu==5.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tornado==6.3.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
djoser==2.2.0
social-auth-app-django==5.2.0